check if value is a digit

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -7,8 +7,8 @@
  */
 void puts_half(char *str)
 {
-	int i = 0;
-	int n;
+	long int i = 0;
+	long int n;
 
 	while (*(str + i) != '\0')
 	{
